Fix wrong cursor placement when editing a task w/ double click
The code in ResizeTreeView.mouseDoubleClickEvent didn't work
for task/project editors, because they're created w/ left padding.
For some reason, the padding is recognized as part of the editor
by mapToParent, but not by the mouse click handler. Because of this,
translated click coordinates were ending up a padding width off
to the right.
Moving the whole editor to the left instead solves the problem,
and has added bonuses:
1. not hiding an icon when editing a task/project
2. stopping an awkward text shift when editing a regular note, caused
by an invalid padding value