Fix subtree displaying bug
Switching between the whole tree and the item's subtree changes
the item's indentation, and, for long enough items, the amount of lines
they need to be drawn. A subtree item height was calculated
in the context of the whole tree, which caused long items to be drawn
w/ dead space underneath.